fastAPI-backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic[email]==2.6.4
pydantic-settings==2.2.1
motor==3.3.2
pymongo==4.6.3
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
httpx==0.27.0
PyJWT==2.8.0
python-dotenv==1.0.1
cloudinary==1.38.0
pytest==8.1.1
pytest-asyncio==0.23.5